feat: Y2021 Day 22
This one is... not good. Not as terrible as it could be, but not
fantastic. The algorithm itself is sufficiently clever, I think, but
it's definitely not the best choice. Part 2's real data takes several
hours, even with PyPy, but it *does* solve it. It's definitely a
candidate for going back to optimise later.