feat: Y2022 Day 17
This one was iffy only in that I had a hard time with the *test* input
on part 2. I figured out cycle detection almost immediately, but the
issue was that the test input never (as far as I found) forms a full
line that allows cutting off the lower part of the rock map, so my loop
detection couldn't find a repeating rock map. Fixed that by only
considering the last thirty lines of input. Feels dirty, but eh.